[FIX] Wargnings at compile time because of comparisons between signed and unsigned values

+++ corewar/src/memory_init/set_params.c
@@ -7,7 +7,7 @@
 
 #include "vm_prototypes.h"
 
-void direct_param(unsigned char const buffer[], vm_t *vm, unsigned int i,
+void direct_param(unsigned char const buffer[], vm_t *vm, int i,
                     param_vm_t *param)
 {
     if (vm->dir_size == 4)
@@ -16,7 +16,7 @@ void direct_param(unsigned char const buffer[], vm_t *vm, unsigned int i,
         param->en.nb = endian_swap_short(retrieve_short(buffer, i));
 }
 
-void check_param(vm_t *vm, elem_t *champ, unsigned int i, param_vm_t *param)
+void check_param(vm_t *vm, elem_t *champ, int i, param_vm_t *param)
 {
     if (param->bit == 0b01) {
         param->param_len = 1;
@@ -30,8 +30,7 @@ void check_param(vm_t *vm, elem_t *champ, unsigned int i, param_vm_t *param)
     }
 }
 
-void fill_params(vm_t *vm, elem_t *champ, unsigned char buffer[],
-                    unsigned int *i)
+void fill_params(vm_t *vm, elem_t *champ, unsigned char buffer[], int *i)
 {
     param_vm_t param = init_param();
 
@@ -39,12 +38,13 @@ void fill_params(vm_t *vm, elem_t *champ, unsigned char buffer[],
     for (unsigned int j = 0; j < 3; j++) {
         param.bit = (buffer[*i] >> ((j * 2 - 6) * -1)) & 0b11;
         if (param.bit != 0b00)
-            check_param(vm, champ, *i + param.len, &param);
+            check_param(vm, champ, *i + (int)param.len, &param);
         else
             break;
-        for (int k = 0; param.param_len != 1 && k < param.param_len; k++)
+        for (unsigned int k = 0; param.param_len != 1 && k < param.param_len;
+        k++)
             vm->mem[champ->address + *i + k + param.len] = param.en.str[k];
         param.len += param.param_len;
     }
-    *i += param.len;
+    *i += (int)param.len;
 }
